ci: Add placeholder for stability tests
The stability tests are currently running as part of the Jenkins CI with
the following setups:
* Container Engines: containerd
* VMMs: QEMU | Cloud Hypervisor | Dragonball
* Snapshotters: overlayfs | devmapper (not for dragonball)

We'll be restricting those tests to be running on LTS version of
containerd.

As it's known due to our GHA limitations, this is just a placeholder and
the tests will actually be added in the interations.

Signed-off-by: Fabiano Fidêncio <fabiano.fidencio@intel.com>