[downstream] runtime: agent: mounts: Mount configfs into the container
configfs is used to get a quote generated, and having this information
available from inside the container (in case the container itself wants
to attest something) is a must.

It's important to note that this is not really needed by Confidential
Containers, as the attestation-agent (which is running on the pod
sandbox VM) has access to this and is able to get the quote generated on
the guest without any kind of issue.  However, this is still reasonable
to have exposed to the container, as it may and will help doing some
simpler tests.

Signed-off-by: Fabiano Fidêncio <fabiano.fidencio@intel.com>